+++ changelog.md
@@ -6,3 +6,7 @@
 - Desarrollo del comando 'copiar' según el requerimiento 4.1
 ## v1.2 [2024-01-13]
 - Desarrollo del comando 'mover' según el requerimiento 4.2
+## v1.3 [2024-01-13]
+- Desarrollo del comando 'renombrar' según el requerimiento 4.3 y Desarrollo del comando 'listar' según requerimirnto 4.4.
+## v1.4 [2024-01-13]
+- Desarrollo del comando 'creardir' según requerimiento 4.5